Move almond datasets and tasks out of `__init__`
